Remove BaseUser class
