--- src/Sulu/Bundle/SecurityBundle/Resources/config/doctrine/BaseRole.orm.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
-                  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                  xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-    <mapped-superclass name="Sulu\Bundle\SecurityBundle\Entity\BaseRole">
-        <id name="id" type="integer" column="id">
-            <generator strategy="AUTO"/>
-        </id>
-
-        <field name="name" type="string" column="name" length="60" unique="true"/>
-        <field name="system" type="string" column="securitySystem" length="60"/>
-
-        <many-to-one field="securityType" target-entity="Sulu\Bundle\SecurityBundle\Entity\SecurityType"
-                     inversed-by="roles">
-            <join-columns>
-                <join-column name="idSecurityTypes" referenced-column-name="id" on-delete="SET NULL" nullable="true"/>
-            </join-columns>
-        </many-to-one>
-
-    </mapped-superclass>
-</doctrine-mapping>
